% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseprices-data.R
\docType{data}
\name{house_prices}
\alias{house_prices}
\title{Prices of houses with attributes}
\format{
A data frame with 546 observations of  12 variables:
}
\source{

}
\usage{
house_prices
}
\description{
A dataset containing the prices and other attributes of
residential houses.
}
\section{Variables}{


\describe{
   \item{price}{num USD}
   \item{lotsize}{int square feet}
   \item{bedrooms}{int number of bedrooms}
   \item{bathrooms}{int number of bathrooms}
   \item{stories}{number of stories in the house}
   \item{driveway}{Factor "no","yes"}
   \item{recreation}{Factor "no","yes" recreation rooms}
   \item{fullbase}{Factor "no","yes" full basement}
   \item{gasheat}{Factor "no","yes"}
   \item{aircon}{Factor "no","yes" air conditioning}
   \item{garage}{int 1 = garage, 0 = no garage}
   \item{prefer}{actor "no","yes"}
}
}

\keyword{datasets}
